Remove old balance implementation
