+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance/BalanceService2.kt
@@ -4,6 +4,7 @@ import io.horizontalsystems.bankwallet.core.*
 import io.horizontalsystems.bankwallet.core.managers.ConnectivityManager
 import io.horizontalsystems.bankwallet.entities.AccountType
 import io.horizontalsystems.bankwallet.entities.Wallet
+import io.horizontalsystems.bankwallet.modules.balance2.BalanceModule2
 import io.horizontalsystems.marketkit.models.CoinPrice
 import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable
@@ -36,8 +37,8 @@ class BalanceService2(
 
     private var hideZeroBalances = false
 
-    private val allBalanceItems = CopyOnWriteArrayList<BalanceModule.BalanceItem>()
-    val balanceItems: List<BalanceModule.BalanceItem>
+    private val allBalanceItems = CopyOnWriteArrayList<BalanceModule2.BalanceItem>()
+    val balanceItems: List<BalanceModule2.BalanceItem>
         get() = if (hideZeroBalances) {
             allBalanceItems.filter { it.balanceData.total > BigDecimal.ZERO }
         } else {
@@ -143,7 +144,7 @@ class BalanceService2(
         val latestRates = xRateRepository.getLatestRates()
 
         val balanceItems = wallets.map { wallet ->
-            BalanceModule.BalanceItem(
+            BalanceModule2.BalanceItem(
                 wallet,
                 networkTypeChecker.isMainNet(wallet),
                 adapterRepository.balanceData(wallet),
