+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance/BalanceSorter.kt
@@ -1,10 +1,11 @@
 package io.horizontalsystems.bankwallet.modules.balance
 
+import io.horizontalsystems.bankwallet.modules.balance2.BalanceModule2
 import java.math.BigDecimal
 
 class BalanceSorter {
 
-    fun sort(items: Iterable<BalanceModule.BalanceItem>, sortType: BalanceSortType): List<BalanceModule.BalanceItem> {
+    fun sort(items: Iterable<BalanceModule2.BalanceItem>, sortType: BalanceSortType): List<BalanceModule2.BalanceItem> {
         return when (sortType) {
             BalanceSortType.Value -> sortByBalance(items)
             BalanceSortType.Name -> items.sortedBy { it.wallet.coin.code }
@@ -12,9 +13,9 @@ class BalanceSorter {
         }
     }
 
-    private fun sortByBalance(items: Iterable<BalanceModule.BalanceItem>): List<BalanceModule.BalanceItem> {
+    private fun sortByBalance(items: Iterable<BalanceModule2.BalanceItem>): List<BalanceModule2.BalanceItem> {
         val comparator =
-                compareByDescending<BalanceModule.BalanceItem> {
+                compareByDescending<BalanceModule2.BalanceItem> {
                     it.balanceData.available > BigDecimal.ZERO
                 }.thenByDescending {
                     it.fiatValue ?: BigDecimal.ZERO > BigDecimal.ZERO
