+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance/BalanceViewItem.kt
@@ -10,6 +10,7 @@ import io.horizontalsystems.bankwallet.core.iconUrl
 import io.horizontalsystems.bankwallet.core.providers.Translator
 import io.horizontalsystems.bankwallet.entities.Wallet
 import io.horizontalsystems.bankwallet.entities.swappable
+import io.horizontalsystems.bankwallet.modules.balance2.BalanceModule2
 import io.horizontalsystems.core.entities.Currency
 import io.horizontalsystems.core.helpers.DateHelper
 import io.horizontalsystems.marketkit.models.CoinPrice
@@ -177,7 +178,7 @@ class BalanceViewItemFactory {
         return DeemedValue(value, deemed, visible)
     }
 
-    fun viewItem(item: BalanceModule.BalanceItem, currency: Currency, expanded: Boolean, hideBalance: Boolean): BalanceViewItem {
+    fun viewItem(item: BalanceModule2.BalanceItem, currency: Currency, expanded: Boolean, hideBalance: Boolean): BalanceViewItem {
         val wallet = item.wallet
         val coin = wallet.coin
         val state = item.state
@@ -215,7 +216,7 @@ class BalanceViewItemFactory {
         )
     }
 
-    fun headerViewItem(items: List<BalanceModule.BalanceItem>, currency: Currency, balanceHidden: Boolean): BalanceHeaderViewItem = when {
+    fun headerViewItem(items: List<BalanceModule2.BalanceItem>, currency: Currency, balanceHidden: Boolean): BalanceHeaderViewItem = when {
         balanceHidden -> BalanceHeaderViewItem("*****", true)
         else -> {
             val total = items.mapNotNull { item ->
