+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance2/BalanceModule2.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.balance
+package io.horizontalsystems.bankwallet.modules.balance2
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -6,14 +6,27 @@ import io.horizontalsystems.bankwallet.core.AdapterState
 import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.core.BalanceData
 import io.horizontalsystems.bankwallet.entities.Wallet
+import io.horizontalsystems.bankwallet.modules.balance.*
 import io.horizontalsystems.marketkit.models.CoinPrice
 
-object BalanceModule {
-
-    class Factory : ViewModelProvider.Factory {
+object BalanceModule2 {
+    class AccountsFactory : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            val activeAccountService = ActiveAccountService(App.accountManager)
+//            val activeAccountService = ActiveAccountService(App.accountManager)
+
+//            val rateAppService = RateAppService(App.rateAppManager)
+
+            return BalanceAccountsViewModel(App.accountManager) as T
+        }
+    }
+
+    class BalanceXxxFactory : ViewModelProvider.Factory {
+        @Suppress("UNCHECKED_CAST")
+        override fun <T : ViewModel> create(modelClass: Class<T>): T {
+//            val activeAccountService = ActiveAccountService(App.accountManager)
+
+//            val rateAppService = RateAppService(App.rateAppManager)
 
             val balanceService2 = BalanceService2(
                 BalanceActiveWalletRepository(App.walletManager, App.accountSettingManager),
@@ -26,14 +39,10 @@ object BalanceModule {
                 App.accountManager
             )
 
-            val rateAppService = RateAppService(App.rateAppManager)
 
             return BalanceViewModel(
                 balanceService2,
-                rateAppService,
-                activeAccountService,
-                BalanceViewItemFactory(),
-                App.appConfigProvider.reportEmail
+                BalanceViewItemFactory()
             ) as T
         }
     }
