+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance2/ui/BalanceForAccount.kt
@@ -15,15 +15,15 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.slideFromBottom
-import io.horizontalsystems.bankwallet.modules.balance.AccountViewItem
-import io.horizontalsystems.bankwallet.modules.balance2.BalanceModule
+import io.horizontalsystems.bankwallet.modules.balance2.AccountViewItem
+import io.horizontalsystems.bankwallet.modules.balance2.BalanceModule2
 import io.horizontalsystems.bankwallet.modules.balance2.BalanceViewModel
 import io.horizontalsystems.bankwallet.modules.manageaccounts.ManageAccountsModule
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 
 @Composable
 fun BalanceForAccount(navController: NavController, accountViewItem: AccountViewItem) {
-    val viewModel = viewModel<BalanceViewModel>(factory = BalanceModule.BalanceXxxFactory())
+    val viewModel = viewModel<BalanceViewModel>(factory = BalanceModule2.BalanceXxxFactory())
 
     Column {
         TopAppBar(
