+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance2/ui/BalanceItemsEmpty.kt
@@ -2,7 +2,7 @@ package io.horizontalsystems.bankwallet.modules.balance2.ui
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import io.horizontalsystems.bankwallet.modules.balance.AccountViewItem
+import io.horizontalsystems.bankwallet.modules.balance2.AccountViewItem
 
 @Composable
 fun BalanceItemsEmpty(navController: NavController, accountViewItem: AccountViewItem) {
